✨ feat(alliance-sidebar): add support for S3 images in sidebar navigation options and refactor interfaces for improved structure
