+++ research-indicators/src/app/shared/interfaces/administration-nav.interface.ts
@@ -1,0 +1,14 @@
+export interface AdministrationNavChild {
+  label: string;
+  link: string;
+  icon?: string;
+  s3Image?: string;
+  hide?: boolean;
+}
+
+export interface AdministrationNavGroup {
+  id: string;
+  label: string;
+  icon: string;
+  children: AdministrationNavChild[];
+}
